WIP: Doesnt update mapping values when no changes

+++ plugins/cardwall/include/OnTop/Config/Command/UpdateMappingFields.class.php
@@ -46,12 +46,24 @@ class Cardwall_OnTop_Config_Command_UpdateMappingFields extends Cardwall_OnTop_C
      */
     private $form_element_factory;
 
-    public function __construct(Tracker $tracker, Cardwall_OnTop_ColumnMappingFieldDao $dao, Cardwall_OnTop_ColumnMappingFieldValueDao $value_dao, TrackerFactory $tracker_factory, Tracker_FormElementFactory $form_element_factory) {
+    /**
+     * @var array
+     */
+    private $existing_mappings;
+
+    public function __construct(
+        Tracker $tracker,
+        Cardwall_OnTop_ColumnMappingFieldDao $dao,
+        Cardwall_OnTop_ColumnMappingFieldValueDao $value_dao,
+        TrackerFactory $tracker_factory,
+        Tracker_FormElementFactory $form_element_factory,
+        array $existing_mappings) {
         parent::__construct($tracker);
         $this->dao                  = $dao;
         $this->value_dao            = $value_dao;
         $this->tracker_factory      = $tracker_factory;
         $this->form_element_factory = $form_element_factory;
+        $this->existing_mappings    = $existing_mappings;
     }
 
     /**
@@ -103,15 +115,42 @@ class Cardwall_OnTop_Config_Command_UpdateMappingFields extends Cardwall_OnTop_C
 
     private function saveValuesMapping(array $mapping_tracker_info, Tracker $mapping_tracker, Tracker_FormElement $field) {
         if (empty($mapping_tracker_info['values']) || !is_array($mapping_tracker_info['values'])) return;
+        $nb_changes = 0;
         foreach ($mapping_tracker_info['values'] as $column_id => $values) {
-            $this->value_dao->deleteAllFieldValues($this->tracker->getId(), $mapping_tracker->getId(), $field->getId(), $column_id);
-            foreach ($values as $value_id) {
-                $this->value_dao->save($this->tracker->getId(), $mapping_tracker->getId(), $field->getId(), (int)$value_id, $column_id);
-                $GLOBALS['Response']->addFeedback('info', 'Mapping values for '. $field->getLabel() .' changed.');
+            if ($this->mappingChanged($mapping_tracker, $column_id, $values)) {
+                $this->value_dao->deleteAllFieldValues($this->tracker->getId(), $mapping_tracker->getId(), $field->getId(), $column_id);
+                foreach ($values as $value_id) {
+                    $nb_changes += $this->value_dao->save($this->tracker->getId(), $mapping_tracker->getId(), $field->getId(), (int)$value_id, $column_id);
+                }
             }
         }
+        $GLOBALS['Response']->addFeedback('info', $nb_changes.' mapping values changed for "'. $field->getLabel().'"');
     }
 
+    private function mappingChanged(Tracker $mapping_tracker, $column_id, array $values) {
+        $no_update_needed = true;
+        if (isset($this->existing_mappings[$mapping_tracker->getId()])) {
+            $value_mappings = $this->existing_mappings[$mapping_tracker->getId()]->getValueMappings();
+            $already_processed = array();
+            foreach ($value_mappings as $value_id => $value_mapping) {
+                $already_processed[$value_id] = true;
+                if ($value_mapping->getColumnId() == $column_id && in_array($value_id, $values)) {
+                    $no_update_needed = $no_update_needed & true;
+                } else {
+                    $no_update_needed = $no_update_needed & false;
+                }
+            }
+            // New values not already mapped;
+            foreach ($values as $value_id) {
+                if (!isset($already_processed[$value_id])) {
+                    $no_update_needed = false;
+                }
+            }
+        } else {
+            $no_update_needed = false;
+        }
+        return !$no_update_needed;
+    }
     /**
      * @return bool
      */
